Get the radius of circles and size of rectangles
Not sure how useful getting the size of a line would be. It's not immediately clear which metric should be used for the length of a line. Either manhattan or chebyshev length will probably be more useful than euclidian length. I'd rather not include all three. Perhaps the size of the rectangle that encloses the line?