Rectangles, circles and lines can only have a stroke width of 1
I just couldn't get QPainter::drawEllipse to produce the output I wanted. If it's not perfect, I won't allow it. If I don't allow circle thickness to be set, I can't allow line and rectangle thickness to be set.

This isn't really that big of a deal anyway. If the user wants thick circles, they can just draw multiple circles. That's what I did with Pixel 1 and it was never a problem! There might be gaps between the circles that the user will have to fill though.