Add StatusMsg
The tool impls file is really starting to bloat. The Tool class really does a lot more than it should. Ideally, it would just manipulate the Cell. However, managing a bunch of parallel class hierarchies for rendering the overlay and updating the status bar is probably worse. I already made a start on that with ToolWidget.

ToolWidget is the widget that will sit in the right side of the status bar and let the user configure the tool. For the brush tool, this means toggling symmetry and changing the thickness. Only some of the tools actually need this widget which is nice.