Convert my font to a TTF
This took me an entire day to do. I wasn't able to find a tool that could do it automatically so I had to use FontForge to manually draw each glyph.

For the first 30 or so glyphs, I was typing in the coordinates of each control point which took a few hours. I then realised that I could mess with the EM size to get snap-to-grid behaviour. I no longer hand to type coordinates so I got the remaining 60 or so done in about half an hour.