Fix editor checkers and compositing
A new checkerboard pattern was used to preserve the 1:1 color ratio when the scale factor is odd. This even preserse the color ratio when the scale factor is 1 so scale 1 looks like scale 2.

The drawing is not converted to a pixmap before rendering. There are some weird alignment issues when drawing pixmaps and images together. Pixmaps must use slightly different rounding or scaling or something.