MEGA COMMIT!
The state of the timeline is nolonger fragmented across a bunch of widgets. The code that modifies the timeline data structure is decoupled from the code that displays it and processes user input.

This has a huge number of advantages and has solved a few problems. The editor widget nolonger has to receive the cells and the their visibilities and combine them together. Since the layers and their visibilities are together in one data structure, the timeline can deal with layer visibilities so that the editor doesn't have to.

Serialization has become much cleaner because all the data is together in one data structure.

Everything has just become much cleaner!