Refactor tools
Tools now have a little bit more control. They interact with the application through signals.

On thing that's a little muddy is the attachCell/detachCell interface. I'm not sure if this abstraction makes sense. Tools might need to know the difference between switching cells and switching tools to provide better status messages. The transform tools are able to provide status messages on attachCell but they don't (to be consistent with the other tools that can't).

Then switching cells, the mouse stays where it is so the position in status bar (written by the brush tool for example) doesn't need to be cleared. When switching tools, the status bar shouldn't contain information from the previous tool so in that case, the status should be cleared. The tools cannot know if the cell is changing or the tool is changing.

I'm not sure if that information is worth providing.