Implement copy, paste, clear in timeline
These functions are really difficult to get right. Manipulating an RLE data structure is a pain. Maybe there's a library for this? I tried looking for one but it's not easy to describe to a search engine.

Implementing this made me wonder how useful these features actually are. How often do you need to copy and clear a range of cells? I guess it's something I'll find out when I start using this tool for real projects. I can recall using the copy and paste feature in Pixel 1 but I don't remember why.