Add down arrow to file input widget
The WidgetRect situation is getting a little messy. It works great for simple situations but it's starting to fall apart.

For simple situations, WidgetRect can store only the necessary information and calculate the rectangles from that. Now we have TextBoxRect which stores a little extra information and provides a slightly different API. This really should be unified into one generic class.

We could have an interface (like a Concept) that has functions for getting rectangles and then a few implementations of that interface that store only the necessary information. That maximises space efficiency but it would require templates everywhere.

We could have a generic WidgetRect that stores the rectangles and then a bunch of factory functions that do all the math. This wastes a little bit of memory but is much more flexible.