Fix launch flicker and workaround positioning issue
There was a noticable flicker when the window was shown because the window was being shown before the widgets were laid out and painted. This was fixed by inserting a few calls to adjustSize and showing the window after the initCanvas signal.

Dock widgets shift their contents up by 1 pixel. I cannot accept this. I tried a bunch of different solutions but this one works the best. I'm just using layouts instead of dock widgets.