Start working on new file format
I started using libpng but I ran into quite a few problems. libpng is not low-level enough for this. libpng is built for writing png images.

Then I started using zlib and it seems to work pretty well. The code that uses deflate is a little tricky to get right though.

I've written very rough draft of the writing side. I still need to work on the reading side and test everything properly.