Start working on sprite deserialization
The current code is able to deserialize part of a file I serialized (AHDR and PLTE) with no errors.

There's so much error checking to do! I'm not really sure how granular the error messages should be. Maybe just saying "File is currupted" would be good enough. If the file is currupted then "canvas width is negative" will be meaningless.