Make it possible to import cells as PNGs
One day, I might make it possible to import part of another sprite into the current sprite. This is a complex operation and I'm not sure if it's necessary. Importing a sequence of PNGs into a sprite might also be possible but I haven't found a need for this either.