Allow the whole layer to become a single row of a sprite sheet
I'm not sure if I like that setting GroupNameMode::empty controls this. I was thinking about something a little more flexible. Instead of just saying "sheet", we could specify a dimension too. Like "sheet_column" or "sheet_row". Then putting the whole layer in a row could be accomplished with GroupNameMode::sheet_column and FrameNameMode::sheet_column. This, of course, allows for 2D sprite sheets as well. But it's a bit tricky to implement. I'd like to avoid a big if-else chain that checks every combination of row and column.